[MAC] Optimisation of mac-sign algo Check .jar file to see if needed to be open to sign files. If not, skip jar

+++ travis/mac-sign.sh
@@ -15,10 +15,14 @@ function signInJar(){
         jar xf "$1" "$f"
 
         if [[ "$f" =~ .*".jar" ]]; then
-            mkdir "_sub" && cd "_sub"
 
-            signInJar "../$f"
-            cd ".." && rm -fr "_sub"
+            # Check if .jar contains necessary to sign files
+            # If not, skip that file and save time
+            if [[ -n $(jar tf "$f" | grep '\.so\|\.dylib\|\.jnilib\|\.jar' | sed -e 's/ //g') ]]; then 
+                mkdir "_sub" && cd "_sub"
+                signInJar "../$f"
+                cd ".." && rm -fr "_sub"
+            fi
         else
             codesign --timestamp --force -s "$MACOS_DEV_ID" -v "$f"
             echo "---"
